requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scipy
altair
Pillow
plotly
scikit-learn
statsmodelsD